fastapi_todo/pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["Jazil Hashmi <jazilhashmi2017@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
pytest = "^8.1.1"
httpx = "^0.27.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
streamlit = "^1.32.2"
bcrypt = "^4.1.2"
status = "^0.2.5"
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
psycopg-binary = "^3.1.19"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
aiokafka = "^0.10.0"
protobuf = ">=3.20,<5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
